requirements.txt
shortuuid~=1.0.1
nanoid~=2.0.0
pydantic~=1.8.1
pymongo~=3.11.3
fastapi~=0.63.0
mongoengine~=0.23.0
motor~=2.4.0
python-decouple